test(server): consolidate the active-run watchdog module tests into 3 files
Each test file pays a new fork, setup load, and transform under the
server project's vitest config. Nine module test files added about
3.85 s of fixed per-file cost against a saving of about 2.58 s from
the Postgres suite trim, so the net wall-clock time went up.

Merge each layer's test files into one file per layer (domain,
application, adapters) to remove the fixed per-file cost while
keeping every test case and assertion.
