+++ server/src/modules/active-run-watchdog/domain/domain.test.ts
@@ -1,5 +1,7 @@
 import { describe, expect, it } from "vitest";
 import { classifySilenceLevel, silenceAgeMs, silenceStartedAt } from "./silence.js";
+import { evaluateSuppression } from "./suppression.js";
+import { isTerminalIssueStatus, shouldFoldTerminalSource } from "./terminal.js";
 
 const SUSPICION_THRESHOLD_MS = 60 * 60 * 1000;
 const CRITICAL_THRESHOLD_MS = 4 * 60 * 60 * 1000;
@@ -173,5 +175,87 @@ describe("classifySilenceLevel", () => {
     },
   ])("classifies: $name", ({ input, expected }) => {
     expect(classifySilenceLevel(input)).toBe(expected);
+  });
+});
+
+describe("evaluateSuppression", () => {
+  it("suppresses a snoozed run until the snooze expires", () => {
+    expect(evaluateSuppression({ snoozedOrContinued: true })).toEqual({
+      suppressed: true,
+      reason: "snoozed",
+    });
+  });
+
+  it("re-arms the watchdog once the continue decision's snooze window has passed", () => {
+    expect(evaluateSuppression({ snoozedOrContinued: false })).toEqual({ suppressed: false });
+  });
+
+  it("suppresses a run permanently after a false-positive decision", () => {
+    expect(evaluateSuppression({ dismissedFalsePositive: true })).toEqual({
+      suppressed: true,
+      reason: "dismissed_false_positive",
+    });
+  });
+
+  it("suppresses a blocked source", () => {
+    expect(evaluateSuppression({ blockedSource: true })).toEqual({
+      suppressed: true,
+      reason: "blocked_source",
+    });
+  });
+
+  it("suppresses a recovery-origin source", () => {
+    expect(evaluateSuppression({ recoveryOriginSource: true })).toEqual({
+      suppressed: true,
+      reason: "recovery_origin_source",
+    });
+  });
+
+  it("is not suppressed when no signal is set", () => {
+    expect(evaluateSuppression({})).toEqual({ suppressed: false });
+  });
+
+  it("checks snoozed before recovery-origin, blocked-source, and dismissed-false-positive", () => {
+    expect(
+      evaluateSuppression({
+        snoozedOrContinued: true,
+        recoveryOriginSource: true,
+        blockedSource: true,
+        dismissedFalsePositive: true,
+      }),
+    ).toEqual({ suppressed: true, reason: "snoozed" });
+  });
+});
+
+describe("isTerminalIssueStatus", () => {
+  it.each([
+    { status: "done", expected: true },
+    { status: "cancelled", expected: true },
+    { status: "in_progress", expected: false },
+    { status: "blocked", expected: false },
+    { status: null, expected: false },
+    { status: undefined, expected: false },
+  ])("$status -> $expected", ({ status, expected }) => {
+    expect(isTerminalIssueStatus(status)).toBe(expected);
+  });
+});
+
+describe("shouldFoldTerminalSource", () => {
+  it("folds a terminal source only with same-run terminal evidence", () => {
+    expect(
+      shouldFoldTerminalSource({ sourceIssueStatus: "done", hasSameRunTerminalEvidence: true }),
+    ).toBe(true);
+  });
+
+  it("does not fold a terminal source without same-run terminal evidence", () => {
+    expect(
+      shouldFoldTerminalSource({ sourceIssueStatus: "done", hasSameRunTerminalEvidence: false }),
+    ).toBe(false);
+  });
+
+  it("does not fold a non-terminal source even with evidence present", () => {
+    expect(
+      shouldFoldTerminalSource({ sourceIssueStatus: "in_progress", hasSameRunTerminalEvidence: true }),
+    ).toBe(false);
   });
 });
